1135589 - move PRIMARY_ID definition
The PRIMARY_ID constant was defined in model.py which also attempted to read
`server.conf` for database connection info. This meant that pulp-admin required
access to `server.conf`.

I moved PRIMARY_ID to pulp.common.constants so it can be imported without
pulling in other code.
